Move Panel from home_screen/home_tab to common.
